Allow float for item costs

+++ includes/parser/plus_format.parser.class.php
@@ -72,7 +72,7 @@ class plus_format extends rli_parser {
 		foreach($xml->items->children() as $xitem) {
 			$cost = (isset($xitem->cost)) ? trim($xitem->cost) : '';
 			$id = (isset($xitem->itemid)) ? trim($xitem->itemid) : '';
-			$data['items'][] = array(trim($xitem->name), trim($xitem->member), (int) $cost, $id, (int) trim($xitem->time));
+			$data['items'][] = array(trim($xitem->name), trim($xitem->member), (float)$cost, $id, (int)trim($xitem->time));
 		}
 		return $data;
 	}
